[DOK-225] style : 폴더 구조 수정
