--- src/components/Bar.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Topbar from "components/Topbar";
-import { Paper } from "@mui/material";
-
-const Bar = () => {
-  return (
-    <Paper>
-      <Topbar />
-    </Paper>
-  );
-};
-
-export default Bar;
