+++ src/components/atoms/Input.js
@@ -1,0 +1,27 @@
+import styled from "styled-components";
+import { Input } from "@mui/joy";
+
+export const InputBoxXS = styled(Input)`
+  width: 210px;
+  height: 50px;
+`;
+
+export const InputBoxS = styled(Input)`
+  width: 520px;
+  height: 50px;
+`;
+
+export const InputBoxM = styled(Input)`
+  width: 820px;
+  height: 50px;
+`;
+
+export const InputBoxL = styled(Input)`
+  width: 1060px;
+  height: 50px;
+`;
+
+export const InputBoxXL = styled(Input)`
+  width: 960px;
+  height: 140px;
+`;
