+++ src/components/organisms/header/Header.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Topbar from "components/Topbar";
-import styled from "@emotion/styled";
+import styled from "styled-components";
 
 const Header = () => {
   return (
