+++ src/components/templete/DefaultLayout.js
@@ -1,6 +1,6 @@
 import Footer from "components/layout/footer/Footer";
 import Header from "components/layout/header/Header";
-import styled from "@emotion/styled";
+import styled from "styled-components";
 
 const DefaultLayout = ({ children }) => {
   return (
