+++ src/components/templete/SidebarLayout.js
@@ -1,7 +1,7 @@
 import { Grid } from "@mui/material";
 import Footer from "components/layout/footer/Footer";
 import Header from "components/layout/header/Header";
-import styled from "@emotion/styled";
+import styled from "styled-components";
 import Sidebar from "components/Sidebar";
 
 const SidebarLayout = ({ children }) => {
