--- src/pages/User/Payment.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Payment = () => {
-  return <div>Payment</div>;
-};
-
-export default Payment;
