--- src/pages/User/ProductList.js
@@ -1,12 +1,0 @@
-import CardGrid from "components/CardGrid";
-import SidebarLayout from "components/layout/SidebarLayout";
-
-const ProductList = () => {
-  return (
-    <SidebarLayout>
-      <CardGrid />
-    </SidebarLayout>
-  );
-};
-
-export default ProductList;
